+++ lvlspy/spcoll.py
@@ -1,5 +1,7 @@
+"""Module to handle a collection of species."""
+
 import os
-import numpy as np
+import sys
 from lxml import etree
 import lvlspy.properties as lp
 import lvlspy.species as ls
@@ -17,17 +19,19 @@ class SpColl(lp.Properties):
     """
 
     def __init__(self, species=None):
+        super().__init__()
         self.properties = {}
         self.spcoll = {}
         if species:
-            for sp in species:
-                self.spcoll[sp.get_name()] = sp
+            for my_species in species:
+                self.spcoll[my_species.get_name()] = my_species
 
     def add_species(self, species):
         """Method to add a species to a collection.
 
         Args:
-            ``species`` (:obj:`lvlspy.species.Species`) The species to be added.
+            ``species`` (:obj:`lvlspy.species.Species`) The species to be
+            added.
 
         Return:
             On successful return, the species has been added.
@@ -40,7 +44,8 @@ class SpColl(lp.Properties):
         """Method to remove a species from a species collection.
 
         Args:
-            ``species`` (:obj:`lvlspy.species.Species`) The species to be removed.
+            ``species`` (:obj:`lvlspy.species.Species`) The species to be
+            removed.
 
         Return:
             On successful return, the species has been removed.
@@ -81,67 +86,62 @@ class SpColl(lp.Properties):
 
         self._add_optional_properties(root, self)
 
-        my_coll = self.get()
+        for species_name, species in self.get().items():
+            my_species = etree.SubElement(root, "species", name=species_name)
 
-        for sp in my_coll:
+            self._add_optional_properties(my_species, species)
 
-            tu_dict = {}
+            xml_levels = etree.SubElement(my_species, "levels")
 
-            for transition in my_coll[sp].get_transitions():
-                e_upper = (transition.get_upper_level()).get_energy()
-
-                if e_upper not in tu_dict:
-                    tu_dict[e_upper] = []
-
-                tu_dict[e_upper].append(transition)
-
-            my_species = etree.SubElement(root, "species", name=sp)
-
-            self._add_optional_properties(my_species, my_coll[sp])
-
-            my_levels = etree.SubElement(my_species, "levels")
-
-            levels = my_coll[sp].get_levels()
-
-            for level in levels:
-                my_level = etree.SubElement(my_levels, "level")
-                self._add_optional_properties(my_level, level)
-                my_level_props = etree.SubElement(my_level, "properties")
-                if units != "keV":
-                    my_energy = etree.SubElement(my_level_props, "energy", units=units)
-                else:
-                    my_energy = etree.SubElement(my_level_props, "energy")
-                my_energy.text = self._get_energy_text(level.get_energy(), units)
-                my_multiplicity = etree.SubElement(my_level_props, "multiplicity")
-                my_multiplicity.text = str(level.get_multiplicity())
-
-                if level.get_energy() in tu_dict:
-
-                    if len(tu_dict[level.get_energy()]) > 0:
-                        my_transitions = etree.SubElement(my_level, "transitions")
-                        for transition in tu_dict[level.get_energy()]:
-                            my_trans = etree.SubElement(my_transitions, "transition")
-                            self._add_optional_properties(my_trans, transition)
-                            lower_level = transition.get_lower_level()
-                            if units != "keV":
-                                my_to_energy = etree.SubElement(
-                                    my_trans, "to_energy", units=units
-                                )
-                            else:
-                                my_to_energy = etree.SubElement(my_trans, "to_energy")
-                            my_to_energy.text = self._get_energy_text(
-                                lower_level.get_energy(), units
-                            )
-                            my_to_multiplicity = etree.SubElement(
-                                my_trans, "to_multiplicity"
-                            )
-                            my_to_multiplicity.text = str(
-                                lower_level.get_multiplicity()
-                            )
-                            my_a = etree.SubElement(my_trans, "a")
-                            my_a.text = str(transition.get_Einstein_A())
+            for level in species.get_levels():
+                self._add_level_to_xml(xml_levels, species, level, units)
 
         xml.write(file, pretty_print=pretty_print)
+
+    def _add_level_to_xml(self, xml_levels, species, level, units):
+        result = etree.SubElement(xml_levels, "level")
+        self._add_optional_properties(result, level)
+        result_props = etree.SubElement(result, "properties")
+        if units != "keV":
+            my_energy = etree.SubElement(result_props, "energy", units=units)
+        else:
+            my_energy = etree.SubElement(result_props, "energy")
+        my_energy.text = self._get_energy_text(level.get_energy(), units)
+        my_multiplicity = etree.SubElement(result_props, "multiplicity")
+        my_multiplicity.text = str(level.get_multiplicity())
+
+        self._add_transitions_to_xml(result, species, level, units)
+
+        return result
+
+    def _add_transitions_to_xml(self, xml_level, species, level, units):
+
+        transitions = species.get_downward_transitions_from_level(level)
+
+        if len(transitions) == 0:
+            return
+
+        xml_transitions = etree.SubElement(xml_level, "transitions")
+
+        for transition in transitions:
+            xml_trans = etree.SubElement(xml_transitions, "transition")
+            self._add_optional_properties(xml_trans, transition)
+            lower_level = transition.get_lower_level()
+            if units != "keV":
+                xml_to_energy = etree.SubElement(
+                    xml_trans, "to_energy", units=units
+                )
+            else:
+                xml_to_energy = etree.SubElement(xml_trans, "to_energy")
+            xml_to_energy.text = self._get_energy_text(
+                lower_level.get_energy(), units
+            )
+            xml_to_multiplicity = etree.SubElement(
+                xml_trans, "to_multiplicity"
+            )
+            xml_to_multiplicity.text = str(lower_level.get_multiplicity())
+            xml_a = etree.SubElement(xml_trans, "a")
+            xml_a.text = str(transition.get_einstein_a())
 
     def _get_energy_text(self, energy, units):
         return str(energy * lv.units_dict[units])
@@ -161,11 +161,15 @@ class SpColl(lp.Properties):
                         )
                     elif len(prop) == 3:
                         my_prop = etree.SubElement(
-                            props, "property", name=prop[0], tag1=prop[1], tag2=prop[2]
+                            props,
+                            "property",
+                            name=prop[0],
+                            tag1=prop[1],
+                            tag2=prop[2],
                         )
                 else:
                     print("Improper property key")
-                    exit()
+                    sys.exit()
 
                 my_prop.text = str(my_props[prop])
 
@@ -195,7 +199,7 @@ class SpColl(lp.Properties):
                     ] = prop.text
                 else:
                     print("Improper keys for property")
-                    exit()
+                    sys.exit()
 
             my_object.update_properties(my_props)
 
@@ -214,11 +218,13 @@ class SpColl(lp.Properties):
         xml = etree.parse(file, parser)
         xml.xinclude()
 
-        schema_file = os.path.join(os.path.dirname(__file__), "xsd_pub/spcoll.xsd")
+        schema_file = os.path.join(
+            os.path.dirname(__file__), "xsd_pub/spcoll.xsd"
+        )
         xmlschema_doc = etree.parse(schema_file)
 
         xml_validator = etree.XMLSchema(xmlschema_doc)
-        xml_validator.assert_(xml)
+        xml_validator.validate(xml)
 
     def update_from_xml(self, file, xpath=""):
         """Method to update a species collection from an XML file.
@@ -242,53 +248,60 @@ class SpColl(lp.Properties):
 
         self._update_optional_properties(spcoll, self)
 
-        el_species = spcoll.xpath("//species" + xpath)
+        for xml_species in spcoll.xpath("//species" + xpath):
+            self.add_species(self._get_species_from_xml(xml_species))
 
-        species = []
+    def _get_species_from_xml(self, xml_species):
+        level_dict = {}
+        result = ls.Species(xml_species.attrib["name"])
+        self._update_optional_properties(xml_species, result)
+        for xml_level in xml_species.xpath(".//level"):
+            new_level = self._get_level_from_xml(xml_level)
+            result.add_level(new_level)
+            level_dict[new_level.get_energy()] = new_level
 
-        for sp in el_species:
-            level_dict = {}
-            my_species = ls.Species(sp.attrib["name"])
-            self._update_optional_properties(sp, my_species)
-            el_level = sp.xpath(".//level")
-            for lev in el_level:
-                props = lev.xpath(".//properties")
-                energy = props[0].xpath(".//energy")
-                multiplicity = props[0].xpath(".//multiplicity")
-                attributes = energy[0].attrib
-                if "units" in attributes:
-                    my_level = lv.Level(
-                        float(energy[0].text),
-                        int(multiplicity[0].text),
-                        units=attributes["units"],
-                    )
-                else:
-                    my_level = lv.Level(
-                        float(energy[0].text), int(multiplicity[0].text)
-                    )
-                self._update_optional_properties(lev, my_level)
+            for xml_trans in xml_level.xpath(".//transition"):
+                trans = self._get_transition_from_xml(
+                    xml_trans, new_level, level_dict
+                )
+                if trans:
+                    result.add_transition(trans)
 
-                level_dict[my_level.get_energy()] = my_level
+        return result
 
-                my_species.add_level(my_level)
+    def _get_level_from_xml(self, xml_level):
+        props = xml_level.xpath(".//properties")
+        energy = props[0].xpath(".//energy")
+        multiplicity = props[0].xpath(".//multiplicity")
+        attributes = energy[0].attrib
+        if "units" in attributes:
+            result = lv.Level(
+                float(energy[0].text),
+                int(multiplicity[0].text),
+                units=attributes["units"],
+            )
+        else:
+            result = lv.Level(float(energy[0].text), int(multiplicity[0].text))
+        self._update_optional_properties(xml_level, result)
 
-                el_trans = lev.xpath(".//transition")
-                for trans in el_trans:
-                    to_energy = trans.xpath(".//to_energy")
-                    to_multiplicity = trans.xpath(".//to_multiplicity")
-                    to_a = trans.xpath(".//a")
+        return result
 
-                    f_to_energy = self._convert_to_keV(to_energy)
-                    if f_to_energy in level_dict:
-                        my_trans = lt.Transition(
-                            my_level, level_dict[f_to_energy], float(to_a[0].text)
-                        )
-                        self._update_optional_properties(trans, my_trans)
-                        my_species.add_transition(my_trans)
+    def _get_transition_from_xml(self, xml_trans, upper_level, level_dict):
+        to_energy = xml_trans.xpath(".//to_energy")
+        to_a = xml_trans.xpath(".//a")
 
-            self.add_species(my_species)
+        f_to_energy = self._convert_to_kev(to_energy)
+        if f_to_energy in level_dict:
+            result = lt.Transition(
+                upper_level,
+                level_dict[f_to_energy],
+                float(to_a[0].text),
+            )
+            self._update_optional_properties(xml_trans, result)
+            return result
+        return None
 
-    def _convert_to_keV(self, energy):
+    def _convert_to_kev(self, energy):
         attributes = energy[0].attrib
         result = float(energy[0].text)
         if "units" in attributes: