+++ lvlspy/transition.py
@@ -1,6 +1,12 @@
+"""Module for handling transitions."""
 import numpy as np
+from gslconsts.consts import (
+    GSL_CONST_CGSM_ELECTRON_VOLT,
+    GSL_CONST_CGS_PLANCKS_CONSTANT_H,
+    GSL_CONST_CGS_SPEED_OF_LIGHT,
+    GSL_CONST_CGSM_BOLTZMANN,
+)
 import lvlspy.properties as lp
-from gslconsts.consts import *
 
 
 class Transition(lp.Properties):
@@ -13,16 +19,18 @@ class Transition(lp.Properties):
         ``lower_level`` (:obj:`lvlspy.level.Level`) The level to which
         there is a spontaneous decay.
 
-        ``Einstein_A`` (:obj:`float`): The Einstein A coefficient (the spontaneous
-        decay rate per second from `upper_level` to `lower_level`).
+        ``einstein_A`` (:obj:`float`): The Einstein A coefficient
+        (the spontaneous decay rate per second from `upper_level` to
+        `lower_level`).
 
     """
 
-    def __init__(self, upper_level, lower_level, Einstein_A):
+    def __init__(self, upper_level, lower_level, einstein_a):
+        super().__init__()
         self.properties = {}
         self.upper_level = upper_level
         self.lower_level = lower_level
-        self.Einstein_A = Einstein_A
+        self.einstein_a = einstein_a
 
     def __eq__(self, other):
         if not isinstance(other, Transition):
@@ -31,7 +39,7 @@ class Transition(lp.Properties):
         return (
             self.upper_level == other.upper_level
             and self.lower_level == other.lower_level
-            and self.Einstein_A == other.Einstein_A
+            and self.einstein_a == other.einstein_a
         )
 
     def get_upper_level(self):
@@ -54,7 +62,7 @@ class Transition(lp.Properties):
 
         return self.lower_level
 
-    def get_Einstein_A(self):
+    def get_einstein_a(self):
         """Method to retrieve the Einstein A coefficient for the transition.
 
         Returns:
@@ -62,68 +70,109 @@ class Transition(lp.Properties):
 
         """
 
-        return self.Einstein_A
+        return self.einstein_a
 
-    def get_Einstein_B_upper_to_lower(self):
+    def update_einstein_a(self, einstein_a):
+        """Method that updates the Einstein A coefficient of a transition.
+
+        Args:
+            ``transition`` (:obj:`lvlspy.transition.Transition`) The
+            transition to be modified.
+
+            ``einstein_A`` (:obj:`float`) The new value for the Einstein A
+            coefficient.
+
+        Returns:
+            On successful return, the transition Einstein A coefficient
+            has been updated.
+        """
+
+        self.einstein_a = einstein_a
+
+    def get_einstein_b_upper_to_lower(self):
         """Method to get the Einstein B coefficient for the upper level
         to lower level transition (induced emission).
 
         Returns:
-            :obj:`float`: The Einstein coefficient in cm\ :sup:`2`
+            :obj:`float`: The Einstein coefficient in cm :sup:`2`
             steradian per erg per s.
 
         """
 
-        nu = self.get_frequency()
-
-        result = self.Einstein_A / self._fnu()
+        result = self.einstein_a / self._fnu()
 
         return result
 
-    def get_Einstein_B_lower_to_upper(self):
+    def get_einstein_b_lower_to_upper(self):
         """Method to get the Einstein B coefficient for the lower level
         to upper level transition (induced absorption).
 
         Returns:
-            :obj:`float`: The Einstein coefficient in cm\ :sup:`2`
+            :obj:`float`: The Einstein coefficient in cm :sup:`2`
             steradian per erg per s.
 
         """
 
-        return self.get_Einstein_B_upper_to_lower() * (
-            self.upper_level.get_multiplicity() / self.lower_level.get_multiplicity()
+        return self.get_einstein_b_upper_to_lower() * (
+            self.upper_level.get_multiplicity()
+            / self.lower_level.get_multiplicity()
         )
 
-    def compute_lower_to_upper_rate(self, T):
-        """Method to compute the total rate for transition from the lower level to
-        upper level.
+    def compute_lower_to_upper_rate(self, temperature, user_func=None):
+        """Method to compute the total rate for transition from the lower
+        level to upper level.
 
         Args:
-            ``T`` (:obj:`float`:) The temperature in K at which to compute
-            the rate.
+            ``temperature`` (:obj:`float`) The temperature in K at which to
+            compute the rate.
+
+            ``user_func`` (optional): A `function
+            <https://docs.python.org/3/library/stdtypes.html#functions>`_
+            that computes the lower level to upper level transition rate.
+            If supplied, the routine will use this function in place of the
+            default one, which computes the rate from the appropriate
+            Einstein coefficient and the blackbody spectrum.
+            The function must take one :obj:`float` argument giving the
+            temperature. Other data can be bound to the function.
 
         Returns:
             :obj:`float`: The rate (per second).
 
         """
 
-        return self.get_Einstein_B_lower_to_upper() * self._bb(T)
+        if user_func:
+            return user_func(temperature)
 
-    def compute_upper_to_lower_rate(self, T):
-        """Method to compute the total rate for transition from the upper level to
-        to lower level.
+        return self.get_einstein_b_lower_to_upper() * self._bb(temperature)
+
+    def compute_upper_to_lower_rate(self, temperature, user_func=None):
+        """Method to compute the total rate for transition from the upper
+        level to to lower level.
 
         Args:
-            ``T`` (:obj:`float`:) The temperature in K at which to compute
-            the rate.
+            ``temperature`` (:obj:`float`) The temperature in K at which to
+            compute the rate.
+
+            ``user_func`` (optional): A `function
+            <https://docs.python.org/3/library/stdtypes.html#functions>`_
+            that computes the upper level to lower level transition rate.
+            If supplied, the routine will use this function in place of the
+            default one, which computes the rate from the appropriate
+            Einstein coefficients and the blackbody spectrum.
+            The function must take one :obj:`float` argument giving the
+            temperature. Other data can be bound to the function.
 
         Returns:
             :obj:`float`: The rate (per second).
 
         """
 
-        return self.get_Einstein_A() + self.get_Einstein_B_upper_to_lower() * self._bb(
-            T
+        if user_func:
+            return user_func(temperature)
+
+        return (
+            self.get_einstein_a()
+            + self.get_einstein_b_upper_to_lower() * self._bb(temperature)
         )
 
     def get_frequency(self):
@@ -134,11 +183,11 @@ class Transition(lp.Properties):
 
         """
 
-        deltaE = self.upper_level.get_energy() - self.lower_level.get_energy()
+        delta_e = self.upper_level.get_energy() - self.lower_level.get_energy()
 
-        deltaE_erg = (1e3) * deltaE * GSL_CONST_CGS_ELECTRON_VOLT
+        delta_e_erg = (1e3) * delta_e * GSL_CONST_CGSM_ELECTRON_VOLT
 
-        return deltaE_erg / GSL_CONST_CGS_PLANCKS_CONSTANT_H
+        return delta_e_erg / GSL_CONST_CGS_PLANCKS_CONSTANT_H
 
     def _fnu(self):
         return (
@@ -148,14 +197,13 @@ class Transition(lp.Properties):
             / np.power(GSL_CONST_CGS_SPEED_OF_LIGHT, 2)
         )
 
-    def _bb(self, T):
-        k_BT = GSL_CONST_CGSM_BOLTZMANN * T
+    def _bb(self, temperature):
+        k_bt = GSL_CONST_CGSM_BOLTZMANN * temperature
 
-        deltaE = self.upper_level.get_energy() - self.lower_level.get_energy()
+        delta_e = self.upper_level.get_energy() - self.lower_level.get_energy()
 
-        x = deltaE * 1.0e3 * GSL_CONST_CGSM_ELECTRON_VOLT / k_BT
+        x_p = delta_e * 1.0e3 * GSL_CONST_CGSM_ELECTRON_VOLT / k_bt
 
-        if x < 500:
-            return self._fnu() / np.expm1(x)
-        else:
-            return self._fnu() * np.exp(-x)
+        if x_p < 500:
+            return self._fnu() / np.expm1(x_p)
+        return self._fnu() * np.exp(-x_p)